SVN revision 10775

+++ examples/certs/pvfs2-ldap-create-dir.sh
@@ -4,8 +4,12 @@
 
 usage ()
 {
-    echo "USAGE: $0 [-h] [-p prefix] [-a admin dn] [-s suffix dn] [-w admin pwd]"
+    echo "USAGE: $0 [-h] [-p prefix] [-c conf dir] [-r run dir] [-d data dir] [-m mod dir] [-a admin dn] [-s suffix dn] [-w admin pwd]"
     echo "    prefix: directory where OpenLDAP is installed, default /usr/local"
+    echo "    config dir: directory containing OpenLDAP server config file (sldap.conf); overrides prefix"
+    echo "    run dir: directory containing pid file etc.; overrides prefix"
+    echo "    data dir: directory containing database files; overrides prefix"
+    echo "    mod dir: directory containing modules (optional); overrides prefix"
     echo "    admin dn: dn of admin user, should be in suffix dn, default cn=admin,{suffix}"
     echo "    suffix dn: base (topmost) dn of tree, default based on hostname, e.g. acme.com => dc=acme,dc=com"
     echo "    admin pwd: admin password, default ldappwd (CHANGE!)"  
@@ -22,11 +26,23 @@ randpw ()
     echo $pass
 }
 
-while getopts "p:a:s:w:" option
+while getopts "p:c:r:d:m:a:s:w:" option
 do
     case $option in
         p)
             prefix=$OPTARG
+        ;;
+        c)
+            confdir=$OPTARG
+        ;;
+        r)
+            rundir=$OPTARG
+        ;;
+        d)
+            datadir=$OPTARG
+        ;;
+        m)
+            moddir=$OPTARG
         ;;
         a)
             admindn=$OPTARG
@@ -49,9 +65,32 @@ if [ ! $prefix ]; then
     prefix=/usr/local
 fi
 
+if [ ! $confdir ]; then
+    confdir=${prefix}/etc/openldap
+fi
+
+if [ ! $rundir ]; then
+    rundir=${prefix}/var/run
+fi
+
+if [ ! $datadir ]; then
+    datadir=${prefix}/var/openldap-data
+fi
+
+if [ ! $moddir ]; then
+    moddir=${prefix}/lib/openldap/modules
+fi
+
 if [ ! $suffix ]; then
     # generate based on hostname
-    hn=`hostname -f`
+    hn=`hostname -f 2> /dev/null`
+    if [ ! $hn ]; then
+         hn=`hostname 2> /dev/null`
+         if [ ! $hn ]; then
+             echo "Error: could not retrieve hostname... exiting"
+             exit 1
+         fi
+    fi
     suffix="dc=${hn/./,dc=}"
 fi
 
@@ -63,8 +102,18 @@ if [ ! $adminpw ]; then
     adminpw="ldappwd"
 fi
 
+# locate slappasswd
+slappasswd=`which slappasswd 2> /dev/null`
+if [ ! $slappasswd ]; then
+    slappasswd=${prefix}/sbin/slappasswd
+    if [ ! -f $slappasswd ]; then
+        echo "Error: could not locate slappasswd; ensure on PATH... exiting"
+        exit 1
+    fi
+fi
+
 # get encrypted root password
-encpwd=`${prefix}/sbin/slappasswd -s $adminpw`
+encpwd=`${slappasswd} -s $adminpw`
 if [ $? -ne 0 ]; then
     echo "Error: could not get password hash... exiting"
     exit 1
@@ -72,7 +121,7 @@ fi
 
 # write the slapd.conf file
 echo -n "Writing slapd.conf... "
-sed "s%__PREFIX__%${prefix}%;s%__ADMINDN__%${admindn}%;s%__ADMINPW__%${encpwd}%;s%__SUFFIX__%${suffix}%" slapd.conf.in > slapd.conf
+sed "s%__PREFIX__%${prefix}%;s%__CONFDIR__%${confdir}%;s%__RUNDIR__%${rundir}%;s%__DATADIR__%${datadir}%;s%__MODDIR__%${moddir}%;s%__ADMINDN__%${admindn}%;s%__ADMINPW__%${encpwd}%;s%__SUFFIX__%${suffix}%" slapd.conf.in > slapd.conf
 if [ $? -eq 0 ]; then
     echo "ok"
 fi
@@ -82,16 +131,29 @@ if [ $? -ne 0 ]; then
    echo "Warning: could not chmod slapd.conf"
 fi
 
+# look for slapd init script
+ldapinit=/etc/init.d/ldap
+if [ ! -f ${ldapinit} ]; then
+    ldapinit=/etc/init.d/slapd
+    if [ ! -f ${ldapinit} ]; then
+        unset ldapinit
+    fi
+fi
+
 # shut down slapd if necessary
 ps -e | grep -q slapd &> /dev/null
 if [ $? -eq 0 ]; then
     echo "Shutting down slapd"
-    killall slapd
+    if [ $ldapinit ]; then
+        $ldapinit stop
+    else
+        killall slapd
+    fi
     sleep 2
 fi
 
 # copy configuration file
-conffile=${prefix}/etc/openldap/slapd.conf
+conffile=${confdir}/slapd.conf
 if [ -f $conffile ]; then
     echo "Backing up existing $conffile to ${conffile}.bak"
     cp -p $conffile ${conffile}.bak
@@ -99,13 +161,18 @@ fi
 
 cp -p slapd.conf $conffile
 
-# locate slapd
-slapd=`which slapd`
-if [ ! $slapd ]; then
-    slapd=${prefix}/libexec/slapd
-    if [ ! -f $slapd ]; then
-        echo "Error: could not locate ${slapd}... exiting"
-        exit 1
+# start slapd using init script
+if [ $ldapinit ]; then
+    $ldapinit start
+else
+    # locate slapd
+    slapd=`which slapd 2> /dev/null`
+    if [ ! $slapd ]; then
+        slapd=${prefix}/libexec/slapd
+        if [ ! -f $slapd ]; then
+            echo "Error: could not locate slapd; ensure on PATH... exiting"
+            exit 1
+        fi
     fi
 fi
